Some updates, not tested
